upgrade view-function from v2 to v3 framework
