--- .changeset/itchy-pigs-train.md
@@ -1,5 +1,0 @@
----
-'@chainlink/view-function-test-adapter': major
----
-
-Initial version of EA
