+++ .changeset/odd-kangaroos-love.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/view-function-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
