--- packages/sources/view-function-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as functionEndpoint } from './function'
