--- packages/sources/view-function-test/src/index.ts
@@ -1,13 +1,0 @@
-import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import { Adapter } from '@chainlink/external-adapter-framework/adapter'
-import { config } from './config'
-import { functionEndpoint } from './endpoint'
-
-export const adapter = new Adapter({
-  defaultEndpoint: functionEndpoint.name,
-  name: 'VIEW-FUNCTION-TEST',
-  config,
-  endpoints: [functionEndpoint],
-})
-
-export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
