--- packages/sources/view-function-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "contract": "0xdAC17F958D2ee523a2206206994597C13D831ec7",
-  "function": "function symbol() view returns (string)"
- }]
-}
