--- packages/sources/view-function-test/test/integration/adapter.test.ts
@@ -1,46 +1,0 @@
-import {
-  TestAdapter,
-  setEnvVariables,
-} from '@chainlink/external-adapter-framework/util/testing-utils'
-import * as nock from 'nock'
-import { mockContractCallResponseSuccess } from './fixtures'
-
-describe('execute', () => {
-  let spy: jest.SpyInstance
-  let testAdapter: TestAdapter
-  let oldEnv: NodeJS.ProcessEnv
-
-  beforeAll(async () => {
-    oldEnv = JSON.parse(JSON.stringify(process.env))
-    process.env.ETHEREUM_RPC_URL = process.env.ETHEREUM_RPC_URL ?? 'http://localhost:8545'
-    const mockDate = new Date('2001-01-01T11:11:11.111Z')
-    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-
-    const adapter = (await import('./../../src')).adapter
-    adapter.rateLimiting = undefined
-    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
-      testAdapter: {} as TestAdapter<never>,
-    })
-  })
-
-  afterAll(async () => {
-    setEnvVariables(oldEnv)
-    await testAdapter.api.close()
-    nock.restore()
-    nock.cleanAll()
-    spy.mockRestore()
-  })
-
-  describe('function endpoint', () => {
-    it('should return success', async () => {
-      const data = {
-        contract: '0xdAC17F958D2ee523a2206206994597C13D831ec7',
-        function: 'function symbol() view returns (string)',
-      }
-      mockContractCallResponseSuccess()
-      const response = await testAdapter.request(data)
-      expect(response.statusCode).toBe(200)
-      expect(response.json()).toMatchSnapshot()
-    })
-  })
-})
