--- packages/sources/view-function-test/test/integration/fixtures.ts
@@ -1,49 +1,0 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import nock from 'nock'
-
-export const mockContractCallResponseSuccess = (): nock.Scope =>
-  nock('http://localhost:8545', {
-    encodedQueryParams: true,
-  })
-    .persist()
-    .post('/', { method: 'eth_chainId', params: [], id: /^\d+$/, jsonrpc: '2.0' })
-    .reply(
-      200,
-      (_, request: AdapterRequest) => ({ jsonrpc: '2.0', id: request['id'], result: '0x1' }),
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
-    .post('/', {
-      method: 'eth_call',
-      params: [{ to: '0xdac17f958d2ee523a2206206994597c13d831ec7', data: '0x95d89b41' }, 'latest'],
-      id: /^\d+$/,
-      jsonrpc: '2.0',
-    })
-    .reply(
-      200,
-      (_, request: AdapterRequest) => ({
-        jsonrpc: '2.0',
-        id: request['id'],
-        result:
-          '0x000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000045553445400000000000000000000000000000000000000000000000000000000',
-      }),
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
-    .persist()
