--- packages/sources/view-function/schemas/env.json
@@ -1,30 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/view-function-adapter.json",
-  "title": "Chainlink External Adapter for View-Function",
-  "description": "External adapter for executing contract function and returning the result",
-  "required": ["ETHEREUM_RPC_URL"],
-  "type": "object",
-  "properties": {
-    "ETHEREUM_RPC_URL": {
-      "type": "string",
-      "description": "RPC URL of a Mainnet ETH node",
-      "required": true
-    },
-    "RPC_URL": {
-      "type": "string",
-      "description": "A fallback RPC URL of a Mainnet ETH node",
-      "required": false
-    },
-    "ETHEREUM_CHAIN_ID": {
-      "required": false,
-      "description": "The chain id to connect to",
-      "type": "string",
-      "default": "1"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
