+++ packages/sources/view-function/src/config/index.ts
@@ -1,34 +1,14 @@
-import { Requester, util } from '@chainlink/ea-bootstrap'
-import { Config as BaseConfig } from '@chainlink/ea-bootstrap'
-import { ethers } from 'ethers'
+import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
 
-export const NAME = 'VIEW_FUNCTION'
-
-export const DEFAULT_ENDPOINT = 'function'
-export const ENV_ETHEREUM_RPC_URL = 'ETHEREUM_RPC_URL'
-export const ENV_FALLBACK_RPC_URL = 'RPC_URL'
-export const ENV_ETHEREUM_CHAIN_ID = 'ETHEREUM_CHAIN_ID'
-export const ENV_FALLBACK_CHAIN_ID = 'CHAIN_ID'
-export const DEFAULT_CHAIN_ID = '1'
-
-export type Config = BaseConfig & {
-  provider: ethers.providers.Provider
-}
-
-export const makeConfig = (prefix?: string): Config => {
-  const rpcURL = util.getRequiredEnvWithFallback(
-    ENV_ETHEREUM_RPC_URL,
-    [ENV_FALLBACK_RPC_URL],
-    prefix,
-  )
-  const chainId =
-    parseInt(
-      util.getEnvWithFallback(ENV_ETHEREUM_CHAIN_ID, [ENV_FALLBACK_CHAIN_ID]) || DEFAULT_CHAIN_ID,
-    ) || util.getEnvWithFallback(ENV_ETHEREUM_CHAIN_ID, [ENV_FALLBACK_CHAIN_ID])
-
-  return {
-    ...Requester.getDefaultConfig(prefix),
-    defaultEndpoint: DEFAULT_ENDPOINT,
-    provider: new ethers.providers.JsonRpcProvider(rpcURL, chainId),
-  }
-}
+export const config = new AdapterConfig({
+  ETHEREUM_RPC_URL: {
+    description: 'RPC URL of a Mainnet ETH node',
+    type: 'string',
+    required: true,
+  },
+  ETHEREUM_CHAIN_ID: {
+    description: 'The chain id to connect to',
+    type: 'number',
+    default: 1,
+  },
+})
