+++ packages/sources/view-function/src/endpoint/function.ts
@@ -1,13 +1,11 @@
-import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { ExecuteWithConfig, InputParameters } from '@chainlink/ea-bootstrap'
-import { Config } from '../config'
-import { utils } from 'ethers'
+import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
+import { InputParameters } from '@chainlink/external-adapter-framework/validation'
+import { config } from '../config'
+import { functionTransport } from '../transport/function'
 
-export const supportedEndpoints = ['function']
-
-export type TInputParameters = { signature: string; address: string; inputParams: string[] }
-export const inputParameters: InputParameters<TInputParameters> = {
+export const inputParameters = new InputParameters({
   signature: {
+    type: 'string',
     aliases: ['function'],
     required: true,
     description:
@@ -20,38 +18,25 @@ export const inputParameters: InputParameters<TInputParameters> = {
     type: 'string',
   },
   inputParams: {
-    required: false,
+    array: true,
     description: 'Array of function parameters in order',
-    type: 'array',
+    type: 'string',
   },
-}
+})
 
-export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
-  const validator = new Validator(request, inputParameters)
-
-  const jobRunID = validator.validated.id
-  const address = validator.validated.data.address
-  const fnSignature = validator.validated.data.signature
-  const params = validator.validated.data.inputParams || []
-
-  const iface = new utils.Interface([fnSignature])
-  const fnName = iface.functions[Object.keys(iface.functions)[0]].name
-
-  const encoded = iface.encodeFunctionData(fnName, [...params])
-
-  const result = await config.provider.call({
-    to: address,
-    data: encoded,
-  })
-
-  const response = {
-    jobRunID,
-    status: 200,
-    statusText: 'OK',
-    headers: {},
-    config: {},
-    data: result,
+export type BaseEndpointTypes = {
+  Parameters: typeof inputParameters.definition
+  Response: {
+    Data: {
+      result: string
+    }
+    Result: string
   }
-
-  return Requester.success(jobRunID, Requester.withResult(response, result))
+  Settings: typeof config.settings
 }
+
+export const endpoint = new AdapterEndpoint({
+  name: 'function',
+  transport: functionTransport,
+  inputParameters,
+})
