+++ packages/sources/view-function/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as FunctionInputParameters } from './function'
-
-export type TInputParameters = FunctionInputParameters
-
-export * as endpoints from './function'
+export { endpoint as functionEndpoint } from './function'
