+++ packages/sources/view-function/test/integration/fixtures.ts
@@ -46,3 +46,4 @@ export const mockContractCallResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
+    .persist()
