+++ packages/tsconfig.json
@@ -561,9 +561,6 @@
       "path": "./sources/view-function"
     },
     {
-      "path": "./sources/view-function-test"
-    },
-    {
       "path": "./sources/wbtc-address-set"
     },
     {
